Pegs Jump AI Puzzle
A variety of puzzles use pegs in a board with a pattern of holes. This Pegs Jump Puzzle is an example of a Pegs puzzle. The puzzle board is comprised of a number of regularly spaced holes. The number of holes that can be discovered differs. Some of the puzzle's holes have pegs in them, while others are left empty. The player progresses by completing a series of jumps. A peg leaps from one peg to the next, landing in an open hole. The peg that was jumped over has now been removed. The goal of the puzzle is to figure out a series of jumps that will result in only one peg on the board and all other holes being empty.